fix: rename component, save token to localStorage, add animation
