+++ src/signin/store/signin.js
@@ -1,7 +1,6 @@
 import Vue from "vue";
 import Vuex from "vuex";
 import axios from "axios";
-//import HTTP from '@/http-common.js';
 Vue.use(Vuex);
 export const HTTP = axios.create({
   baseURL: `http://intern-2021.southeastasia.cloudapp.azure.com:4000/api/v1`,
@@ -11,9 +10,8 @@ export const namespaced = true;
 
 export const mutations = {
   SET_USER_DATA(state, userData) {
-    state.user = userData;
-    localStorage.setItem("user", JSON.stringify(userData));
-    HTTP.defaults.headers.common["Authorization"] = `Bearer ${userData.token}`;
+    const token = userData.token;
+    localStorage.setItem("token", token);
   },
 };
