+++ src/views/AuthFooter.vue
