+++ src/views/AuthPage.vue
@@ -1,18 +1,18 @@
 <template>
   <div class="bg-main fill-height-flex">
     <v-layout column class="fill-height">
-      <HeaderLayout class="header" />
+      <AuthHeader class="header" />
       <router-view class="content" />
-      <FooterLayout class="footer" />
+      <AuthFooter class="footer" />
     </v-layout>
   </div>
 </template>
 
 <script>
-import HeaderLayout from "@/views/HeaderLayout.vue";
-import FooterLayout from "@/views/FooterLayout.vue";
+import AuthHeader from "@/views/AuthHeader.vue";
+import AuthFooter from "@/views/AuthFooter.vue";
 export default {
-  components: { HeaderLayout, FooterLayout },
+  components: { AuthHeader, AuthFooter },
 };
 </script>
